requirements.txt
pymupdf
sentence-transformers 
qdrant-client
google-generativeai
python-dotenv
streamlit 
fastapi
pydantic
uvicorn[standard]
python-multipart